Raksha-Women-Saftey-main/Raksha-Women-Saftey-main/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.26.0
geopy==2.2.0
SQLAlchemy==1.4.23
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
click==8.0.1
Babel==2.9.1
pytz==2021.1
urllib3==1.26.6
certifi==2021.5.30
idna==3.2
chardet==4.0.0
PyJWT==2.1.0 